Adding Tokens/Permissons to user

+++ springboot-first-app/src/main/java/app/filerepository/delete/controller/Delete.java
@@ -1,5 +1,6 @@
 package app.filerepository.delete.controller;
 
+import app.config.TokenProvider;
 import app.response.ResponseMessage;
 import app.filerepository.delete.service.intf.IDeleteService;
 import org.slf4j.Logger;
@@ -9,10 +10,16 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.servlet.http.HttpServletRequest;
+
+import static app.userauth.role.RoleEnum.ADMIN;
+
 @RestController
 public class Delete {
     @Autowired
     private IDeleteService deleteService;
+    @Autowired
+    TokenProvider jwtTokenUtils;
     private static final Logger logger = LoggerFactory.getLogger(Delete.class);
 
     @DeleteMapping("/delete/{fileId}")
@@ -22,8 +29,12 @@ public class Delete {
     }
 
     @DeleteMapping("/deleteAll")
-    public ResponseMessage deleteAllFiles(){
+    public ResponseMessage deleteAllFiles(HttpServletRequest request){
+        String role =   jwtTokenUtils.getRole(request);
         logger.info("deleteAllFiles Response created ");
-        return deleteService.deleteAllFiles();
+        if(role.equalsIgnoreCase(ADMIN.name())){
+            return deleteService.deleteAllFiles();
+        }
+        return ResponseMessage.getInstance("you don't have permission to delete files");
     }
 }
